Make FutureStateChange a top-level class
